Treat default methods like locally declared methods
Prior to this commit, a non-overridden interface default method would
incorrectly be considered overridden by a locally declared method that
overloads the signature for the default method.

This commit addresses this issue by treating default methods as if they
were locally declared methods.

In addition, this commit improves the documentation for
ReflectionSupport and overhauls the internals of ReflectionUtils in the
following ways.

- removal of dead code
- removal of superfluous precondition checks
- consistent error messages for violated preconditions
- enforcement of immutable return types where applicable
- improved diagnostics for failed parameter type resolution
- stable sorting of methods declared in interfaces

Issue: #976