Read ConfigurationParameters from properties file in classpath
Prior to this commit, configuration parameters could only be set via
the LauncherDiscoveryRequestBuilder or JVM system properties; however,
neither of these options were robust solutions for repeatable builds
with regards to configuration parameters such as JUnit Jupiter's
'junit.jupiter.testinstance.lifecycle.default' that alter the manner in
which tests are executed.

This commit addresses this issue by introducing support for supplying
_configuration parameters_ via a file named 'junit-platform.properties'
in the root of the class path. Consult the new 'Configuration
Parameters' section of the User Guide for further details.

Issue: #1003